# docs/swagger.yaml
basePath: /
definitions:
...
  dto.AvatarInfo:
    properties:
      avatar_url:
        example: https://pps.meow.net/...
        type: string
      jid:
        example: 5511999999999@s.meow.net
        type: string
      phone:
        example: "5511999999999"
        type: string
      picture_id:
        example: pic_123
        type: string
      timestamp:
        example: "2023-01-01T12:00:00Z"
        type: string
    type: object
...
  dto.CheckContactRequest:
    properties:
      phones:
        example:
        - '["5511999999999"'
        - ' "5511888888888"]'
        items:
          type: string
        type: array
    required:
    - phones
    type: object
...
  dto.ContactCheckResult:
    properties:
      is_in_meow:
        example: true
        type: boolean
      jid:
        example: 5511999999999@s.meow.net
        type: string
      query:
        example: "5511999999999"
        type: string
      verified_name:
        example: João Silva
        type: string
    type: object
  dto.ContactData:
    properties:
      action:
        example: check_contacts
        type: string
      avatar:
        $ref: '#/definitions/dto.AvatarInfo'
      check_results:
        items:
          $ref: '#/definitions/dto.ContactCheckResult'
        type: array
      contact_infos:
        items:
          $ref: '#/definitions/dto.ContactInfo'
        type: array
      status:
        example: success
        type: string
      timestamp:
        example: "2023-01-01T00:00:00Z"
        type: string
    type: object
  dto.ContactErrorResponse:
    properties:
      code:
        example: INVALID_PHONE
        type: string
      details:
        example: Phone number must include country code
        type: string
      message:
        example: Invalid phone number format
        type: string
    type: object
  dto.ContactInfo:
    properties:
      avatar:
        example: https://...
        type: string
      business_name:
        example: Empresa João
        type: string
      device_count:
        example: 2
        type: integer
      display_name:
        example: João Silva
        type: string
      is_blocked:
        example: false
        type: boolean
      is_muted:
        example: false
        type: boolean
      jid:
        example: 5511999999999@s.meow.net
        type: string
      name:
        example: João Silva
        type: string
      notify:
        example: João
        type: string
      phone:
        example: "5511999999999"
        type: string
      picture_id:
        example: pic_123
        type: string
      push_name:
        example: João
        type: string
      status:
        example: Disponível
        type: string
      verified_name:
        example: João Silva Empresa
        type: string
    type: object
...
  dto.ContactResponse:
    properties:
      code:
        type: integer
      data:
        $ref: '#/definitions/dto.ContactData'
      error:
        $ref: '#/definitions/dto.ContactErrorResponse'
      success:
        type: boolean
    type: object
  dto.ContactsData:
    properties:
      action:
        example: get_contacts
        type: string
      contacts:
        items:
          $ref: '#/definitions/dto.ContactInfo'
        type: array
      count:
        example: 10
        type: integer
      status:
        example: success
        type: string
      timestamp:
        example: "2023-01-01T00:00:00Z"
        type: string
    type: object
...
  dto.ContactsResponse:
    properties:
      code:
        type: integer
      data:
        $ref: '#/definitions/dto.ContactsData'
      error:
        $ref: '#/definitions/dto.ContactErrorResponse'
      success:
        type: boolean
    type: object
...
  dto.GetAvatarRequest:
    properties:
      phone:
        example: "5511999999999"
        type: string
    required:
    - phone
    type: object
...
  dto.GetContactInfoRequest:
    properties:
      phones:
        example:
        - '["5511999999999"'
        - ' "5511888888888"]'
        items:
          type: string
        type: array
    required:
    - phones
    type: object
...
  dto.SetContactPresenceRequest:
    properties:
      state:
        example: available
        type: string
    required:
    - state
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@meow.com
    name: meow API Support
    url: https://github.com/your-username/meow
  description: A meow API server built with Go, inspired by meow
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: meow meow API
  version: "1.0"
paths:
...
  /session/{sessionId}/contacts/avatar:
    post:
      consumes:
      - application/json
      description: Get contact's profile picture/avatar
      parameters:
      - description: Session ID
        in: path
        name: sessionId
        required: true
        type: string
      - description: Get avatar request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.GetAvatarRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ContactResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ContactResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ContactResponse'
      security:
      - ApiKeyAuth: []
      summary: Get contact avatar
      tags:
      - Contacts
  /session/{sessionId}/contacts/check:
    post:
      consumes:
      - application/json
      description: Check if phone numbers are registered on meow
      parameters:
      - description: Session ID
        in: path
        name: sessionId
        required: true
        type: string
      - description: Check contact request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CheckContactRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ContactResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ContactResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ContactResponse'
      security:
      - ApiKeyAuth: []
      summary: Check contacts on meow
      tags:
      - Contacts
  /session/{sessionId}/contacts/info:
    post:
      consumes:
      - application/json
      description: Get detailed information about contacts
      parameters:
      - description: Session ID
        in: path
        name: sessionId
        required: true
        type: string
      - description: Get contact info request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.GetContactInfoRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ContactResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ContactResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ContactResponse'
      security:
      - ApiKeyAuth: []
      summary: Get contact information
      tags:
      - Contacts
  /session/{sessionId}/contacts/list:
    get:
      consumes:
      - application/json
      description: Get all contacts from contact's meow
      parameters:
      - description: Session ID
        in: path
        name: sessionId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ContactsResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ContactResponse'
      security:
      - ApiKeyAuth: []
      summary: Get contacts
      tags:
      - Contacts
  /session/{sessionId}/contacts/presence:
    post:
      consumes:
      - application/json
      description: Set global contact presence (available/unavailable)
      parameters:
      - description: Session ID
        in: path
        name: sessionId
        required: true
        type: string
      - description: Set presence request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.SetContactPresenceRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ContactResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ContactResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ContactResponse'
      security:
      - ApiKeyAuth: []
      summary: Set contact presence
      tags:
      - Contacts
...
schemes:
- http
- https
securityDefinitions:
  ApiKeyAuth:
    description: 'API Key authentication. Simply provide your API key directly: "YOUR_API_KEY".
      The system automatically detects if it''s a Global API Key (can access all sessions
      and session management) or a Session-specific API Key (can only access the specific
      session it belongs to).'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...